Add files via upload
A pdf merger and parser. Merge together similar PDF's then search for where certain words are. The script will display text around the word in the console.  This does need editing in order to export the text to a word doc. 

I built this for work as I deal with a lot of SOPs.